+++ src/callbacks/postPickupInit.ts
@@ -1,8 +1,8 @@
 import { ModCallback, PickupVariant } from "isaac-typescript-definitions";
+import * as fastAngels from "../classes/features/optional/bosses/FastAngels";
 import { PickupVariantCustom } from "../enums/PickupVariantCustom";
 import * as flipCustom from "../features/items/flipCustom";
 import * as seededGBBug from "../features/mandatory/seededGBBug";
-import * as fastAngels from "../features/optional/bosses/fastAngels";
 import * as fastKrampus from "../features/optional/bosses/fastKrampus";
 import * as scaredHeart from "../features/optional/graphics/scaredHeart";
 import * as stickyNickel from "../features/optional/graphics/stickyNickel";